Set revision.txt on both `git push` and `deis pull`.
When using `git push` to push code to deis there is no git repository
and therefore we cannot use git command to get the latest SHA. Instead
we can use GIT_SHA env variable, which is automatically set by Deis.

When using `deis pull` we can still rely on git.

We moved the code that sets revision.txt to run-docker.sh to set
revision.txt based on the environment.