Use python:3.3-slim docker image.
Interstingly when using python:3.4-slim `docker push` fails with the
following message `FATA[0000] Error: Status 400 trying to push
repository mozillamffos/masterfirefoxos-pr: "Access denied:
103b75c0109601b4f73067f9786025fe208c073feba5256ef9adc5cb53ec171d is a
private image" `

Searching reveals that this may be related to the docker version we use
in our build server (docker 1.5) and upgrading to docker 1.6 would fix
the issue. Unfortunatelly we are tied to version 1.5 because 1.6 is not
supported by Deis yet.